50 Projects in 50 Days - Testimonial Box Switcher
Day 47 project. This involved setting up a simple container with a testimonial, and user information (image, name, role, etc). A progress bar was animated using CSS to move from left to right across the container in 10 seconds; once finished, JS was used to add a new testimonial to the page.